add presentation delegate to separate scoped component
