+++ app/src/main/java/org/stepic/droid/di/course_list/CourseListComponent.kt
@@ -9,7 +9,6 @@ import org.stepic.droid.ui.fragments.CourseListFragmentBase
 import org.stepic.droid.ui.fragments.CourseSearchFragment
 import org.stepic.droid.ui.fragments.FastContinueFragment
 import org.stepik.android.view.injection.course.CourseDataModule
-import org.stepik.android.view.injection.course_continue.CourseContinueModule
 import org.stepik.android.view.injection.course_payments.CoursePaymentsDataModule
 import org.stepik.android.view.injection.last_step.LastStepDataModule
 import org.stepik.android.view.injection.personal_deadlines.PersonalDeadlinesDataModule
@@ -29,8 +28,7 @@ import org.stepik.android.view.injection.unit.UnitDataModule
     ProgressDataModule::class,
 
     CourseDataModule::class,
-    PersonalDeadlinesDataModule::class,
-    CourseContinueModule::class
+    PersonalDeadlinesDataModule::class
 ])
 interface CourseListComponent {
