+++ app/src/main/java/org/stepik/android/view/injection/course/CoursePresentationScope.kt
@@ -1,0 +1,7 @@
+package org.stepik.android.view.injection.course
+
+import javax.inject.Scope
+
+@Scope
+@Retention(AnnotationRetention.RUNTIME)
+annotation class CoursePresentationScope
