+++ app/src/main/java/org/stepik/android/view/injection/profile/ProfileComponent.kt
@@ -5,7 +5,6 @@ import dagger.Subcomponent
 import org.stepik.android.view.achievement.ui.fragment.AchievementsListFragment
 import org.stepik.android.view.injection.certificate.CertificateDataModule
 import org.stepik.android.view.injection.course.CourseDataModule
-import org.stepik.android.view.injection.course_continue.CourseContinueModule
 import org.stepik.android.view.injection.course_payments.CoursePaymentsDataModule
 import org.stepik.android.view.injection.last_step.LastStepDataModule
 import org.stepik.android.view.injection.social_profile.SocialProfileDataModule
@@ -29,7 +28,6 @@ import org.stepik.android.view.profile_notification.ui.fragment.ProfileNotificat
     CoursePaymentsDataModule::class,
     LastStepDataModule::class,
     ProfileDataModule::class,
-    CourseContinueModule::class,
     UserDataModule::class,
     UserActivityDataModule::class,
     SocialProfileDataModule::class,
